+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+- update kaioretry.decorator module documentation
+- update kaioretry.context module documentation
+
 ## [0.9.0] - 2023-02-25
 
 ### Changed